+++ classes/beverage.py
@@ -1,4 +1,6 @@
+from .beverageExceptions import *
 import copy
+
 class Beverage:
     """Napój używany w automacie"""
     def __init__(self, name):
@@ -6,21 +8,6 @@ class Beverage:
     def get_name(self):
         """Zwraca nazwę napoju"""
         return self.__beverageName
-
-class BeveragesAmountException(Exception):
-    """Reaguje na ilość poniżej 0"""
-    def __init__(self, amount):
-        super().__init__(f"Podano za małą ilość: {amount}")
-
-class BeveragesPriceException(Exception):
-    """Reaguje na cenę mniejszą od 1 gr"""
-    def __init__(self, price):
-        super().__init__(f"Podano złą cenę: {price}")
-    
-class OutOfStorageException(Exception):
-    """Nie ma napojów"""
-    def __init__(self):
-        super().__init__("Nie ma już więcej sztuk")
 
 class BeveragesInfo:
     """Zawiera informacje o cenie, ilości i instancji"""
